crawler/requirements.txt
requests
cassandra-driver